permit disabling random offset, and extract common query parts
I've disabled "flaky" from the tests since I'm not even sure that
helps - is there a case where the second or third try succeeds? Or do
the failures tend to happen because either expected 2 got 4 or
expected 4 got 3, or expected 2 got 0, and consistently we get the
same (unexpected) count each time. My gut feeling is the difference
here is between the state of the database (which is the same between
runs) and retrying the same setup back to back will produce the same
problematic results.